Polish guided CLI installation
