+++ desktop/src/shared/api/tauri.ts
@@ -1,7 +1,5 @@
 import { invoke as tauriInvoke } from "@tauri-apps/api/core";
 import type {
-  AcpAuthMethod,
-  AcpAuthMethodsResult,
   AddChannelMembersInput,
   AddChannelMembersResult,
   BackendProviderCandidate,
@@ -41,7 +39,6 @@ import type {
   AcpRuntimeCatalogEntry,
   AuthStatus,
   CommandAvailability,
-  ConnectAcpRuntimeResult,
   InstallRuntimeResult,
   GitBashPrerequisite,
   OpenDmInput,
@@ -236,24 +233,6 @@ export type RawAcpRuntimeCatalogEntry = {
   /** Tagged union with snake_case status values — same shape as `AuthStatus`. */
   auth_status: AuthStatus;
   login_hint?: string;
-};
-
-export type RawAcpAuthMethod = {
-  id: string;
-  name: string;
-  description?: string | null;
-  type?: string | null;
-  args?: string[];
-  command?: string[];
-  _meta?: unknown;
-};
-
-export type RawAcpAuthMethodsResult = {
-  methods: RawAcpAuthMethod[];
-};
-
-export type RawConnectAcpRuntimeResult = {
-  launched: boolean;
 };
 
 export type RawInstallStepResult = {
@@ -936,18 +915,6 @@ function fromRawAcpRuntimeCatalogEntry(
   };
 }
 
-function fromRawAcpAuthMethod(method: RawAcpAuthMethod): AcpAuthMethod {
-  return {
-    id: method.id,
-    name: method.name,
-    description: method.description ?? null,
-    type: method.type ?? null,
-    args: method.args ?? [],
-    command: method.command ?? [],
-    meta: method._meta ?? null,
-  };
-}
-
 function fromRawInstallRuntimeResult(
   raw: RawInstallRuntimeResult,
 ): InstallRuntimeResult {
@@ -1118,25 +1085,6 @@ export async function discoverGitBashPrerequisite(): Promise<GitBashPrerequisite
       installHint: prerequisite.install_hint,
     }
   );
-}
-
-export async function discoverAcpAuthMethods(
-  runtimeId: string,
-): Promise<AcpAuthMethodsResult> {
-  const raw = await invokeTauri<RawAcpAuthMethodsResult>(
-    "discover_acp_auth_methods",
-    { runtimeId },
-  );
-  return { methods: raw.methods.map(fromRawAcpAuthMethod) };
-}
-
-export async function connectAcpRuntime(
-  runtimeId: string,
-  methodId: string,
-): Promise<ConnectAcpRuntimeResult> {
-  return invokeTauri<RawConnectAcpRuntimeResult>("connect_acp_runtime", {
-    request: { runtimeId, methodId },
-  });
 }
 
 export async function discoverAcpRuntimes(): Promise<AcpRuntimeCatalogEntry[]> {
